Change timer implementation.
setInterval is not guaranteed to actually be called at the value specified.  When the page is in the background or minimized timers are slowed down on modern browsers.

This implementation sets an endTime using Date() and the display instead just shows the remaining time, regardless of how slow the setInterval callback is called.

Fixes: #11